fix(lbac-3055): revue RQTH
