+++ server/src/jobs/simpleJobDefinitions.ts
@@ -3,11 +3,11 @@ import { updateDiplomeMetier } from "@/jobs/diplomesMetiers/updateDiplomesMetier
 import { updateRomesForDomainesMetiers } from "@/jobs/domainesMetiers/updateRomesForDomainesMetiers"
 import { sendMiseEnRelation } from "@/jobs/miseEnRelation/sendMiseEnRelation"
 import {
-  processMeteojob,
   processAtlas,
-  processViteUnEmploi,
+  processMeteojob,
   processNosTalentsNosEmplois,
   processToulouseMetropole,
+  processViteUnEmploi,
 } from "@/jobs/offrePartenaire/clever-connect/processCleverConnect"
 import { importRecruteursLbaFromComputedToJobsPartners } from "@/jobs/offrePartenaire/fillComputedRecruteursLba"
 import { classifyFranceTravailJobs } from "@/jobs/offrePartenaire/france-travail/classifyJobsFranceTravail"
@@ -25,7 +25,6 @@ import { renvoiMailCreationCompte } from "@/jobs/oneTimeJob/renvoiMailCreationCo
 import { exportJobsToS3V2 } from "@/jobs/partenaireExport/exportJobsToS3V2"
 import { exportJobsToFranceTravail } from "@/jobs/partenaireExport/exportToFranceTravail"
 import { repriseEnvoiEmailsPRDV } from "@/jobs/rdv/repriseEnvoiPRDV"
-import { sendMailEngagementHandicap } from "@/jobs/recruiters/sendMailEngagementHandicap"
 import { processScheduledRecruiterIntentions } from "@/services/application.service"
 import { generateSitemap } from "@/services/sitemap.service"
 
@@ -371,9 +370,5 @@ export const simpleJobDefinitions: SimpleJobDefinition[] = [
   {
     fct: removeBrevoContacts,
     description: "Anonymise les contacts Brevo dont la date de creation est supérieure à 2 ans",
-  },
-  {
-    fct: sendMailEngagementHandicap,
-    description: "Envoie les mails de sensibilisation à l'handi-engagement",
   },
 ]
