+++ server/src/services/etablissement.service.ts
@@ -30,6 +30,7 @@ import { getStaticFilePath } from "@/common/utils/getStaticFilePath"
 import { getHttpClient } from "@/common/utils/httpUtils"
 import { getDbCollection } from "@/common/utils/mongodbUtils"
 import { userWithAccountToUserForToken } from "@/security/accessTokenService"
+import { sendEngagementHandicapEmailIfNeeded } from "@/services/handiEngagement.service"
 import { getUserWithAccountByEmail, isUserEmailChecked } from "@/services/userWithAccount.service"
 
 import { isEmailFromPrivateCompany, isEmailSameDomain } from "../common/utils/mailUtils"
@@ -569,7 +570,7 @@ export const entrepriseOnboardingWorkflow = {
     const opcoResult = await updateEntrepriseOpco(siret, { opco, idcc: parseInt(idcc ?? "") || null })
 
     let validated = false
-    const managingUser = await createOrganizationUser({
+    const { user: managingUser, role } = await createOrganizationUser({
       userFields: { first_name, last_name, phone: phone ?? "", origin, email: formatedEmail },
       is_email_checked: false,
       organization: { type: ENTREPRISE, entreprise },
@@ -604,6 +605,8 @@ export const entrepriseOnboardingWorkflow = {
       },
       managingUser._id.toString()
     )
+
+    await sendEngagementHandicapEmailIfNeeded(managingUser, role._id, entreprise)
 
     return { formulaire, user: managingUser, validated }
   },
